requirements.txt
torch>=2.4.0
torchvision
transformers>=4.45.0
accelerate>=0.34.0
bitsandbytes>=0.43.3

qwen-vl-utils

sentence-transformers
numpy

fastapi
uvicorn
python-multipart
pydantic-settings

Pillow
python-dotenv